Fixing Department From Float to Double

+++ src/main/kotlin/com/prasetia/erp/model/department/DepartmentBudgetDetail.kt
@@ -11,9 +11,9 @@ class DepartmentBudgetDetail(
         val line_id:Long?,
         val code:String?,
         val budget_item_view:String?,
-        val nilai_budget:Long?,
-        val realisasi_budget:Long?,
-        val persent_budget: Float?
+        val nilai_budget:Double?,
+        val realisasi_budget:Double?,
+        val persent_budget: Double?
 ){
-    constructor(): this(0,0,0,"","", 0,0,0f)
+    constructor(): this(0,0,0,"","", 0.0,0.0,0.0)
 }
