+++ src/main/kotlin/com/prasetia/erp/pojo/department/DepartmentSummaryData.kt
@@ -6,9 +6,9 @@ import javax.persistence.Id
 class DepartmentSummaryData(
         val id:Long,
         val periode:Long?,
-        val nilai_budget:Long?,
-        val realisasi_budget:Long?,
-        val persent_budget:Float?
+        val nilai_budget:Double?,
+        val realisasi_budget:Double?,
+        val persent_budget:Double?
 ){
-    constructor(): this(0,0, 0,0,0f)
+    constructor(): this(0,0, 0.0,0.0,0.0)
 }
